fixed all register controller bugs

+++ src/utils/cloudinary.js
@@ -9,21 +9,26 @@ cloudinary.config({
 
 const uploadOnCloudinary = async (localFilePath) => {
     try {
-        if(!localFilePath) return null
+        if (!localFilePath) return null
         //upload the file on cloudinary
-        const response = await cloudinary.uploader.upload
-        (localFilePath, {
+        const response = await cloudinary.uploader.upload(localFilePath, {
             resource_type: "auto"
         })
-        //file has been uploaded successfully
-        console.log("File is uploaded on cloudinary", response.url);
+        // file has been uploaded successfull
+        //console.log("file is uploaded on cloudinary ", response.url);
+        // fs.unlinkSync(localFilePath)
+        // return response;
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }
         return response;
+
     } catch (error) {
-        fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed 
-        
+       if (localFilePath && fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        } // remove the locally saved temporary file as the upload operation got failed
+        return null;
     }
 }
-    
-    // console.log(uploadResult); 
 
 export {uploadOnCloudinary}
